requirements.txt
Django==5.2
djangorestframework==3.16
drf-yasg==1.21.10